pyproject.toml
[project]
name = "local-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Carberry",email = "alex.carberry@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "watchdog (>=6.0.0,<7.0.0)",
    "prompt-toolkit (>=3.0.48,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "numpy (>=2.2.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
